+++ lib/livingstyleguide/command_line_interface.rb
@@ -10,6 +10,7 @@ module LivingStyleGuide
   class CommandLineInterface < Thor
 
     desc 'compile input output', 'Compiles the living style guide to HTML.'
+    option :stdout, type: :boolean
     def compile(input, output = nil)
       output = input.sub(/(\.html)?\.lsg$/, '.html') if output.nil?
       if defined?(Compass)
@@ -17,8 +18,12 @@ module LivingStyleGuide
         output = output.sub /^#{Compass.configuration.sass_dir}/, Compass.configuration.css_dir
       end
       html = Tilt.new(input).render
-      File.write output, html
-      puts "Successfully generated a living style guide at #{output}."
+      if options[:stdout]
+        puts html
+      else
+        File.write output, html
+        puts "Successfully generated a living style guide at #{output}."
+      end
     end
 
   end